[SWM-173] Refactor : change project structure
