+++ app/script/script.py
@@ -1,0 +1,5 @@
+class Script:
+    def __init__(self):
+        self.raw_script = ''
+        self.text = ''
+        self.token_count = 0
